Make starting and stop the server more fault tolerant.

+++ sh/stop.sh
@@ -1,3 +1,7 @@
 #!/bin/sh
 
-killall nodejs
+if pgrep nodejs >/dev/null; then
+  killall nodejs
+else
+  echo "No nodejs process found, continuing..."
+fi
